+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## [unpublished]
 
-- Changed
+- Added
   - Added `DOC504` and a config option
     `--should-declare-assert-error-if-assert-statement-exists`. If this option
     is True and a function has an `assert` statement, an `AssertError`
     declaration is required in the docstring. Otherwise `DOC504` is raised.
     (This changes the behavior introduced in v0.6.1.)
+  - Added a new config option `--ignore-private-args` (default to `False`)
 
 ## [0.6.2] - 2025-02-17